Split `Graph` from `RoutingTableView`
