+++ chain/network/benches/graph.rs
@@ -2,7 +2,7 @@
 extern crate bencher;
 
 use bencher::Bencher;
-use near_network::routing::Graph;
+use near_network::routing::graph::Graph;
 
 use near_network::test_utils::random_peer_id;
