+++ chain/network/src/routing/mod.rs
@@ -1,4 +1,5 @@
 pub(crate) mod edge_validator_actor;
+pub mod graph;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub(crate) mod ibf;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
@@ -7,18 +8,14 @@ pub(crate) mod ibf_peer_set;
 pub(crate) mod ibf_set;
 pub(crate) mod network_protocol;
 mod route_back_cache;
-#[allow(clippy::module_inception)]
-pub(crate) mod routing;
-pub(crate) mod routing_table_actor;
-
 pub use crate::network_protocol::{Edge, EdgeState, PartialEdgeInfo, SimpleEdge};
 #[cfg(feature = "test_features")]
 pub use crate::private_actix::GetRoutingTableResult;
+pub(crate) mod routing_table_actor;
+pub mod routing_table_view;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub use crate::routing::ibf_peer_set::SlotMapId;
 #[cfg(feature = "protocol_feature_routing_exchange_algorithm")]
 pub use crate::routing::ibf_set::IbfSet;
-pub use crate::routing::routing::{
-    Graph, RoutingTableView, DELETE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME,
-};
-pub use crate::routing::routing_table_actor::{start_routing_table_actor, Prune};
+
+pub use routing_table_actor::start_routing_table_actor;
