+++ chain/network/src/tests/cache_edges.rs
@@ -1,5 +1,6 @@
-use crate::network_protocol::Edge;
-use crate::routing::{EdgeState, Prune, DELETE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME};
+use crate::network_protocol::{Edge, EdgeState};
+use crate::routing::routing_table_actor::Prune;
+use crate::routing::routing_table_view::{DELETE_PEERS_AFTER_TIME, SAVE_PEERS_MAX_TIME};
 use crate::test_utils::random_peer_id;
 use crate::RoutingTableActor;
 use actix::System;
